processColors moved to shared function
