+++ tests/app.test.js
@@ -5,7 +5,7 @@ const request = require("supertest");
 const { client } = require("../shared/redis-client.js");
 const { ACCESS_TOKEN_SECRET } = require("../shared/constants.js");
 const { truncTables, insertRows } = require("./config/setup-jest.js");
-const { fetchProductData } = require("../fetch/fetch-products.js");
+const fetchProductData = require("../fetch/fetch-products.js");
 const { subscriberInit, subscriber } = require("../shared/subscriber-init.js");
 const {
   insertData,
